Fix unwanted results when merging config arrays
The native array_merge_recursive() function was used to merge the config
arrays. But this resulted in not supported configs when both default and
project specific config file contained the same string key. The native
function combines 2 values with same key into an array.

Added custom method to combine config arrays who overwrite the value
of the first array with the value from the second.